feat(shannon): genai functions supported
